alteração de tipo dos id

+++ src/main/kotlin/com/stockcore/dto/MovimentacaoCreateDTO.kt
@@ -3,7 +3,7 @@ package com.stockcore.dto
 import java.time.LocalDateTime
 
 data class MovimentacaoCreateDTO(
-    val produtoId: Long,
+    val produtoId: Int,
     val quantidade: Int,
     val tipoMovimentacao: String,
     val dataHora: LocalDateTime,
