+++ src/main/kotlin/com/stockcore/dto/ProdutoCreateDTO.kt
@@ -2,7 +2,7 @@ package com.stockcore.dto
 
 data class ProdutoCreateDTO(
     val nome: String,
-    val tipoId: Long,
+    val tipoId: Int,
     val quantidade: Int,
     val descricao: String?
 )
