+++ src/main/kotlin/com/stockcore/dto/ProdutoUpdateDTO.kt
@@ -2,7 +2,7 @@ package com.stockcore.dto
 
 data class ProdutoUpdateDTO(
     val nome: String,
-    val tipoId: Long,
+    val tipoId: Int,
     val quantidade: Int,
     val descricao: String?
 )
