+++ src/main/kotlin/com/stockcore/model/TipoProduto.kt
@@ -8,7 +8,7 @@ data class TipoProduto(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idTipoProduto")
-    val idTipoProduto: Long = 0,
+    val idTipoProduto: Int = 0,
 
     @Column(nullable = false, unique = true)
     val nomeTipo: String
