+++ src/main/kotlin/com/stockcore/model/Usuario.kt
@@ -8,7 +8,7 @@ data class Usuario(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idUsuario")
-    val idUsuario: Long = 0,
+    val idUsuario: Int = 0,
 
     @Column(nullable = false, unique = true)
     val nome: String,
